Added Mana Bar, fixed targeting bug
Added in a new bar for mana, re-arranged the xp and health bars to fit it. It's kind of not great, we should probably get some dedicated art on this. Also fixed a bug where canFire was defaulting to true, meaning that cancelling a targeting panel at the beginning of an attack would still eat your turn.